Stop using circleci
It was getting too difficult to find a docker image and compiler
which had support for C++17 and C style non trivially initilizers,
which gcc8.2 did not. Fighting with circleci is not what I'm
interested in solving in this project for now, so I have put
setting up CI for a later time to focus more on the actual
project.