Change size of reference to be 1
References sizes are inexplicably 4 bytes wide in java, which I did not know before.

This ends up jumping through lots of hoops to satisfy this requirement, most anoyingly, pointers to objects are now stored in a global map. This creates a lot of anoying problems and is acutally significantly slower, but the alternative of creating (or linking against) a custom allocator where I take either the top or bottom 32 bits of the address space for it would be too big of a burden. This is a really terrible design choice to require that references be only 4 bytes. Now every object access, not just creation, requires the aquisition of a mutex. Ridiculous.