Add 'filerunner' test infrastructure
This patch adds new kinds of tests for testing output of the JVM. It also adds the 'filerunner_test' target which should run all the tests.

'filerunner' is a go program heavily inspired by LLVM-lit. It was written in Go because I haven't used the language much and didn't want to write this in C++ nor in Python because I don't like Python. It isn't well written but I wanted to get something quick out, I think the tests are useful regardless. Unfortunately this adds a dependency on a Go compiler, which is a significant burden, but it is just for a small subset of tests.

For examples on how it works check test/filerunner/FilerunnerTest.java.